adjust code structure
